% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestPictures.R
\name{bestPictures}
\alias{bestPictures}
\title{Kopiere das beste Bild eines jeden Ereignisses in einen export_Ordner
(hierbei wird das Bild verkleinert und ein Overlay hinzugefügt)}
\usage{
bestPictures(file_path, mask = TRUE, scale = TRUE, size = "x480")
}
\arguments{
\item{file_path}{character, Pfad zu dem Speicherort für die Bilder und dem Hauptordner der Ereignisse}

\item{mask}{boolean, gibt an, ob die FVA-Bildmaske hinzugefügt werden soll}

\item{scale}{boolean, wenn TRUE werden Bilder entsprechend dem size-Parameter angepasst (s.u.)}

\item{size}{character, gibt die Bildhöhe an. "x480" ändert die Bildhöhe auf 480 Pixel, der aspect ratio wird beibehalten}
}
\value{
nothing
}
\description{
Kopiere das beste Bild eines jeden Ereignisses in einen export_Ordner
(hierbei wird das Bild verkleinert und ein Overlay hinzugefügt)
}
