% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurPerson.R
\name{blurPerson}
\alias{blurPerson}
\title{Augewählte Fotos werden verpixelt und ggf. verkleinert}
\usage{
blurPerson(
  file_path,
  scale = TRUE,
  size = 480,
  column,
  selection = c("Mensch", "Mensch u Hund", ...)
)
}
\arguments{
\item{file_path}{Pfad zum Ordner der Session. Beispielsweise "N:/Projekte/Gruenbrueckenmonitor_1708/Daten/Feldarbeit/Fotofallenmonitoring/session_07".}

\item{scale}{Gibt an, ob das Foto verkleinert werden soll. Default = TRUE}

\item{size}{Größe des Fotos, wenn diese verändert werden soll. Default = "x480"}

\item{column}{Name der Spalte, in der die Klassifizierung steht. Beispielsweise "Tierart_1".}

\item{selection}{Character Vector mit den Klassen, die verpixelt werden sollen. Beispielsweise c("Mensch","Mensch u Hund").}
}
\value{
Die Fotos werden am Speicherort überschrieben.
}
\description{
Augewählte Fotos werden verpixelt und ggf. verkleinert
}
