% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFoldernames.R
\name{checkFoldernames}
\alias{checkFoldernames}
\title{Ersetzt Ordnernamen durch tatsächlichen Ordnernamen}
\usage{
checkFoldernames(importTables_list)
}
\arguments{
\item{importTables_list}{list, Liste die mit DMCr2::importTables() und DMCr2::processImportTables() erstellt wurde}
}
\description{
Ersetzt Ordnernamen durch tatsächlichen Ordnernamen
}
