% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLocationnames.R
\name{checkLocationnames}
\alias{checkLocationnames}
\title{Idenfitiziert falsche bzw noch nicht in der Datenbank abgelegte Standortbezeichnungen}
\usage{
checkLocationnames(importTables_list)
}
\arguments{
\item{importTables_list}{list, Liste die mit DMCr2::importTables() und DMCr2::processImportTables() erstellt wurde}
}
\value{
data.table, Tabelle mit fraglichen Standortbezeichnungen
}
\description{
Idenfitiziert falsche bzw noch nicht in der Datenbank abgelegte Standortbezeichnungen
}
