% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpeciesCorr.R
\name{checkSpeciesCorr}
\alias{checkSpeciesCorr}
\title{Korrigiert Tierartenbezeichnungen gemäß der manuellen Eingabe}
\usage{
checkSpeciesCorr(importTables_list, falsche_tierarten_df)
}
\arguments{
\item{importTables_list}{list, Liste die mit DMCr2::importTables() und DMCr2::processImportTables() erstellt wurde}

\item{falsche_tierarten_df}{data.frame, Datentabelle mit alten und korrigierten Tierartenbezeichnungen}
}
\value{
list, Liste mit Datentabellen mit korrigierten Tiernamen
}
\description{
Korrigiert Tierartenbezeichnungen gemäß der manuellen Eingabe
}
