% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpeciesWrong.R
\name{checkSpeciesWrong}
\alias{checkSpeciesWrong}
\title{Korrigiert Tierartenbezeichnungen gemäß der manuellen Eingabe}
\usage{
checkSpeciesWrong(importTables_list, speciesDB_vector)
}
\arguments{
\item{importTables_list}{list, Liste die mit DMCr2::importTables() und DMCr2::processImportTables() erstellt wurde}

\item{speciesDB_vector}{character vector, Tierartbezeichnungen in der Datenbank}
}
\value{
data.frame, Tabelle mit falschen, zu korrigierenden bzw überprüfenden Tierartbezeichnungen
}
\description{
Korrigiert Tierartenbezeichnungen gemäß der manuellen Eingabe
}
