% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condaEnvironment.R
\name{condaEnvironment}
\alias{condaEnvironment}
\title{Erstellung einer Conda-Umgebung für den MegaDetector}
\usage{
condaEnvironment()
}
\value{
conda env
}
\description{
Erstellung einer Conda-Umgebung für den MegaDetector
}
