% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnectionUpload.R
\name{dbConnectionUpload}
\alias{dbConnectionUpload}
\title{Eine Probe-Verbindung zur FVA-Fotofallendatenbank mit Schreibrechten, um das Passwort zu 端berpr端fen.}
\usage{
dbConnectionUpload(psw)
}
\value{
Formal class PostgreSQLConnection
}
\description{
Eine Probe-Verbindung zur FVA-Fotofallendatenbank mit Schreibrechten, um das Passwort zu 端berpr端fen.
}
