% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMCr2-package.R
\docType{package}
\name{DMCr2-package}
\alias{DMCr2}
\alias{DMCr2-package}
\title{DMCr2: Data Management of Camera Trap Data}
\description{
This package facilitates the transfer of FFM2-classified camera trap images and camera trap locations between a PostgreSQL-Database and a local system.
}
\author{
\strong{Maintainer}: Anja Müller-Meißner \email{anja.mueller-meissner@forst.bwl.de}

Authors:
\itemize{
  \item Max Kröschel
  \item Alexander Milles
}

}
\keyword{internal}
