% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDB.R
\name{downloadDB}
\alias{downloadDB}
\title{Ruft Daten von der FVA-Fotofallendatenbank ab.}
\usage{
downloadDB(data, species, project)
}
\arguments{
\item{data}{Character, Bestimmt welche Art von Daten (data = "ereignisse", data = "bilder" oder data = "standorte")  abgefragt werden sollen. Es ist nur ein Wert erblaubt.}

\item{species}{Character vector mit Arten, von denen Daten abgefragt werden sollen. Der Wert "alle Arten" ruft Daten für alle in der Datenbank vorhandene Arten ab. Falls data = "standorte" werden nur Standorte abgefragt, an denen eine bestimmte Tierart in fotofallen.ereignisse vorkommt (tierart_1 oder tierart_2).}

\item{project}{Character vector, Auswahl eines Projekts, mehrerer oder aller Projekte, für das Daten abgefragt werden sollen. Der Wert "alle Projekte" ruft Daten für alle in der Datenbank vorhandene Projekte ab.}
}
\value{
data.frame
}
\description{
Ruft Daten von der FVA-Fotofallendatenbank ab.
}
