% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Ruft Daten von der FVA-Fotofallendatenbank ab.}
\usage{
extract(data, species, project)
}
\arguments{
\item{data}{Bestimmt welche Art von Daten abgefragt werden sollen. Es ist nur ein Wert erblaubt.}

\item{species}{Character vector mit Arten, von denen Daten abgefragt werden sollen. Der Wert "alle Arten" ruft Daten für alle in der Datenbank vorhandene Arten ab.}

\item{project}{Auswahl eines Projekts oder alle, für das Daten abgefragt werden sollen. Der Wert "alle Projekte" ruft Daten für alle in der Datenbank vorhandene Projekte ab.}
}
\value{
data.frame
}
\description{
Ruft Daten von der FVA-Fotofallendatenbank ab.
}
