% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPictures.R
\name{extractPictures}
\alias{extractPictures}
\title{Ruft Bilder aus dem Ordnerverzeichnis anhand der Datenbankinfo ab}
\usage{
extractPictures(
  bilder_table,
  save_path,
  scale = TRUE,
  size = "x480",
  tags = NULL,
  test_mode = FALSE
)
}
\arguments{
\item{bilder_table}{data.table, Datentabelle mit Information zu den Pfaden zu den Bildern (i.d.R. erstellt mit downloadDB("bilder", "alle Arten", "alle Projekte"))}

\item{save_path}{character, Pfad zu dem Speicherort für die Bilder}

\item{scale}{boolean, wenn TRUE werden Bilder entsprechend dem size-Parameter angepasst (s.u.)}

\item{size}{character, gibt die Bildhöhe an. "x480" ändert die Bildhöhe auf 480 Pixel, der aspect ratio wird beibehalten}

\item{tags}{character vector, gibt die tags an, die in den Bildern eingetragen werden sollen}

\item{test_mode}{boolean, nur für Testzwecke zu nutzen}
}
\value{
data.frame
}
\description{
Ruft Bilder aus dem Ordnerverzeichnis anhand der Datenbankinfo ab
}
