% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDBStandorte.R
\name{mapDBStandorte}
\alias{mapDBStandorte}
\title{Plottet eine Karte von Baden-W체rttemberg mit den ausgw채hlten Standorten von Fotofallenkameras.}
\usage{
mapDBStandorte()
}
\value{
recordedplot (als Liste)
}
\description{
Plottet eine Karte von Baden-W체rttemberg mit den ausgw채hlten Standorten von Fotofallenkameras.
}
