% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megaDetector.R
\name{megaDetector}
\alias{megaDetector}
\title{Ausführung des MegaDetectors}
\usage{
megaDetector(file_path, run = c("batch", "simple"))
}
\arguments{
\item{file_path}{Pfad zum Ordner der Session (wenn run = "batch") oder zum Foto (wenn run = "simple"). Beispielsweise "N:/Projekte/Gruenbrueckenmonitor_1708/Daten/Feldarbeit/Fotofallenmonitoring/session_07" bzw. ".../session_07/Fotos/Böblingen/GBB_ff01_210805/I_00008a.JPG".}

\item{run}{run = "batch" führt den MegaDetector für mehrere Fotos aus, run = "simple" für ein einzelnes.}
}
\value{
JSON-Datei (wenn run = "batch) oder ein Foto als JPG mit bounding box (wenn run = "simple").
}
\description{
Ausführung des MegaDetectors
}
