% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMegaDetectorblurPerson.R
\name{outputMegaDetectorblurPerson}
\alias{outputMegaDetectorblurPerson}
\title{Fotos mit "Personen" (Klassifikation vom MegaDetector) drauf werden verpixelt + Thumbnail aktualisiert}
\usage{
outputMegaDetectorblurPerson(file_path)
}
\arguments{
\item{file_path}{Pfad zur Session}
}
\description{
Fotos mit "Personen" (Klassifikation vom MegaDetector) drauf werden verpixelt + Thumbnail aktualisiert
}
