% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMegaDetectorcsv.R
\name{outputMegaDetectorcsv}
\alias{outputMegaDetectorcsv}
\title{Umwandlung der MegaDetector-Datei (JSON) in eine csv-Tabelle}
\usage{
outputMegaDetectorcsv(file_path)
}
\arguments{
\item{file_path}{Pfad zur Session}
}
\description{
Umwandlung der MegaDetector-Datei (JSON) in eine csv-Tabelle
}
