% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMegaDetectorFolder.R
\name{outputMegaDetectorFolder}
\alias{outputMegaDetectorFolder}
\title{Fotos mit Detection des MegaDetectors werden in einen separaten Ordner verschoben}
\usage{
outputMegaDetectorFolder(file_path)
}
\arguments{
\item{file_path}{Dateipfad zum Foto (.jpg)}
}
\description{
Fotos mit Detection des MegaDetectors werden in einen separaten Ordner verschoben
}
