% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMegaDetectorRename.R
\name{outputMegaDetectorRename}
\alias{outputMegaDetectorRename}
\title{Bereits existierende MegaDetector-Datei wird umbenannt, damit sie nicht überschrieben wird}
\usage{
outputMegaDetectorRename(file_path)
}
\arguments{
\item{file_path}{Pfad zur Datei "Megadetector_output.json"}
}
\description{
Bereits existierende MegaDetector-Datei wird umbenannt, damit sie nicht überschrieben wird
}
