% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMegaDetectorSeparate.R
\name{outputMegaDetectorSeparate}
\alias{outputMegaDetectorSeparate}
\title{Durch den MegaDetector klassifizierte Fotos werden nach Kategorie der Detectino in separate Ordner kopiert.}
\usage{
outputMegaDetectorSeparate(file_path)
}
\arguments{
\item{file_path}{}
}
\description{
Durch den MegaDetector klassifizierte Fotos werden nach Kategorie der Detectino in separate Ordner kopiert.
}
