% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processImportTables.R
\name{processImportTables}
\alias{processImportTables}
\title{Bereinigt die mittels importTables geladenen Tabellen}
\usage{
processImportTables(daten_ereignis_list)
}
\arguments{
\item{daten_ereignis_list}{list, enthält die beiden data.tables daten und ereignis (output von importTables)}
}
\value{
data.table
}
\description{
Bereinigt die mittels importTables geladenen Tabellen
}
