% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFiles_getVolumes.R
\name{shinyFiles_getVolumes}
\alias{shinyFiles_getVolumes}
\title{Debugged copy of shinyFiles::getVolumes()
This function is intended as an input to the roots parameter in fileGetter() and shinyFileChoose(). It returns a function that returns a named vector of available volumes on the system. This construction makes it dynamic so that a shinyFiles instance reflects new volumes as they get added (e.g. usb drives). The function takes a single argument giving names of volumes the developer wants removed from the return value.}
\usage{
shinyFiles_getVolumes(exclude = NA)
}
\arguments{
\item{exclude}{A vector of volume names to be excluded from the return value}
}
\value{
character vector of available volumes
}
\description{
Debugged copy of shinyFiles::getVolumes()
This function is intended as an input to the roots parameter in fileGetter() and shinyFileChoose(). It returns a function that returns a named vector of available volumes on the system. This construction makes it dynamic so that a shinyFiles instance reflects new volumes as they get added (e.g. usb drives). The function takes a single argument giving names of volumes the developer wants removed from the return value.
}
