% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesDB.R
\name{speciesDB}
\alias{speciesDB}
\title{Gibt alle Tierarten in fotofallen.ereignisse in der Datenbank an}
\usage{
speciesDB()
}
\value{
character vector
}
\description{
Gibt alle Tierarten in fotofallen.ereignisse in der Datenbank an
}
\examples{
#' speciesDB()

}
