% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_upload.R
\name{upload}
\alias{upload}
\title{vermutlich veraltete Funktion, wird aktuell noch belassen
Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)}
\usage{
upload()
}
\description{
vermutlich veraltete Funktion, wird aktuell noch belassen
Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)
}
