% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_uploadFinal.R
\name{uploadFinal}
\alias{uploadFinal}
\title{vermutlich veraltete Funktion, wird aktuell noch belassen
Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)}
\usage{
uploadFinal(user, psw)
}
\description{
vermutlich veraltete Funktion, wird aktuell noch belassen
Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)
}
