% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadStandorte.R
\name{uploadStandorte}
\alias{uploadStandorte}
\title{Standorte der Session werden in die FVA-Fotofallendatenbank (fotofallen.fotofallen_standorte_import) hochgeladen (PostgreSQL)}
\usage{
uploadStandorte(upload_sf)
}
\arguments{
\item{upload_sf}{simple feature Objekt, Fotofallenstandorte mit korrigierten bzw. überprüften Spaltenbezeichnungen}
}
\description{
Standorte der Session werden in die FVA-Fotofallendatenbank (fotofallen.fotofallen_standorte_import) hochgeladen (PostgreSQL)
}
