% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPictures.R
\name{extractPictures}
\alias{extractPictures}
\title{Ruft Bilder aus dem Ordnerverzeichnis anhand der Datenbankinfo ab}
\usage{
extractPictures(bilder_table, save_path, scale = TRUE, size = "x480", tags)
}
\arguments{
\item{bilder_table}{data.table, Datentabelle mit Information zu den Pfaden zu den Bildern (i.d.R. erstellt mit downloadDB("bilder", "alle Arten", "alle Projekte"))}

\item{save_path}{character, Pfad zu dem Speicherort für die Bilder}

\item{scale}{character, gibt die Bildhöhe an. "x480" ändert die Bildhöhe auf 480 Pixel, der aspect ratio wird beibehalten}

\item{tags}{character vector, gibt die tags an, die in den Bildern eingetragen werden sollen}
}
\value{
data.frame
}
\description{
Ruft Bilder aus dem Ordnerverzeichnis anhand der Datenbankinfo ab
}
