% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFinal.R
\name{uploadFinal}
\alias{uploadFinal}
\title{Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)}
\usage{
uploadFinal(user, psw)
}
\description{
Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)
}
