% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadImport.R
\name{uploadImport}
\alias{uploadImport}
\title{Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)}
\usage{
uploadImport(importTables_list)
}
\arguments{
\item{importTables_list}{list, Liste mit Daten und Ereignissen als data.tables nach Durchführung der Korrekturfunktionen}
}
\value{
Nothing
}
\description{
Datentabellen (daten und ereignis) der Session werden in die FVA-Fotofallendatenbank hochgeladen (PostgreSQL)
}
